+++ app/templates/settings/personal.php
@@ -1,0 +1,4 @@
+<?php
+script('learning', 'learning-personal-settings');
+?>
+<div id="learning-personal-settings"></div>